# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Evaluation Service API
  version: 1.2.0
  description: |
    Full API contract for the current backend implementation (Phases 1-5).
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/evaluations:
    post:
      tags: [Evaluations]
      summary: Submit evaluation
      description: |
        Submits and scores an evaluation. When `features.enable-step-windows=true`,
        the assignment's step window policy is enforced (enabled step, open/close window,
        optional late window).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitEvaluationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
  /api/v1/evaluations/validate-submit:
    post:
      tags: [Evaluations]
      summary: Validate submission payload before submit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitEvaluationRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionValidationResponse'
  /api/v1/evaluations/{id}:
    get:
      tags: [Evaluations]
      summary: Get evaluation
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
    put:
      tags: [Evaluations]
      summary: Save/update evaluation draft
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvaluationRequest'
      responses:
        '200':
          description: Updated evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
  /api/v1/evaluations/{id}/receipt:
    get:
      tags: [Evaluations]
      summary: Get submission receipt by evaluation id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Submission receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionReceiptResponse'
  /api/v1/evaluations/{id}/admin-detail:
    get:
      tags: [Evaluations]
      summary: Admin submission detail projection
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Admin submission detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubmissionDetailResponse'
  /api/v1/evaluations/campaign/{campaignId}:
    get:
      tags: [Evaluations]
      summary: List evaluations by campaign
      parameters:
        - in: path
          name: campaignId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: Evaluation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationResponse'
  /api/v1/evaluations/evaluatee/{evaluateeId}:
    get:
      tags: [Evaluations]
      summary: List evaluations by evaluatee
      parameters:
        - in: path
          name: evaluateeId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: Evaluation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationResponse'
  /api/v1/evaluations/{id}/flag:
    post:
      tags: [Evaluations]
      summary: Flag evaluation
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Flagged
  /api/v1/evaluations/{id}/invalidate:
    post:
      tags: [Evaluations]
      summary: Invalidate evaluation
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Invalidated
  /api/v1/evaluations/{id}/reopen:
    post:
      tags: [Evaluations]
      summary: Reopen evaluation for revision (admin)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Reopened
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
    SubmitEvaluationRequest:
      type: object
      required: [campaignId, assignmentId, evaluatorId, evaluateeId, templateId, answers]
      properties:
        campaignId: { type: string }
        assignmentId: { type: string }
        evaluatorId: { type: string }
        evaluateeId: { type: string }
        templateId: { type: string }
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerRequest'
    UpdateEvaluationRequest:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerRequest'
    AnswerRequest:
      type: object
      properties:
        questionId: { type: string }
        value: {}
        selectedOptions:
          type: array
          items: { type: string }
        textResponse: { type: string }
        metadata:
          $ref: '#/components/schemas/AnyObject'
    EvaluationResponse:
      type: object
      properties:
        id: { type: string }
        campaignId: { type: string }
        assignmentId: { type: string }
        evaluatorId: { type: string }
        evaluateeId: { type: string }
        templateId: { type: string }
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerResponse'
        status:
          type: string
          enum: [DRAFT, SUBMITTED, SCORING, COMPLETED, FLAGGED, INVALIDATED]
        totalScore: { type: number, format: double, nullable: true }
        answerCount: { type: integer }
        createdAt: { type: string, format: date-time }
        submittedAt: { type: string, format: date-time, nullable: true }
    SubmissionValidationResponse:
      type: object
      properties:
        valid: { type: boolean }
        issueCount: { type: integer }
        issues:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionValidationIssue'
    SubmissionValidationIssue:
      type: object
      properties:
        questionId: { type: string, nullable: true }
        code: { type: string }
        message: { type: string }
    SubmissionReceiptResponse:
      type: object
      properties:
        evaluationId: { type: string }
        campaignId: { type: string }
        assignmentId: { type: string }
        evaluatorId: { type: string }
        evaluateeId: { type: string }
        status:
          type: string
          enum: [DRAFT, SUBMITTED, SCORING, COMPLETED, FLAGGED, INVALIDATED]
        submittedAt: { type: string, format: date-time, nullable: true }
        totalScore: { type: number, format: double, nullable: true }
    AnswerResponse:
      type: object
      properties:
        id: { type: string }
        questionId: { type: string }
        value: { type: string }
        selectedOptions:
          type: array
          items: { type: string }
        textResponse: { type: string }
        metadata:
          $ref: '#/components/schemas/AnyObject'
...
    AdminSubmissionDetailResponse:
      type: object
      properties:
        evaluationId: { type: string }
        evaluationStatus:
          type: string
          enum: [DRAFT, SUBMITTED, SCORING, COMPLETED, FLAGGED, INVALIDATED]
        submittedAt: { type: string, format: date-time, nullable: true }
        totalScore: { type: number, format: double, nullable: true }
        campaignId: { type: string }
        campaignName: { type: string }
        campaignStatus:
          type: string
          enum: [DRAFT, SCHEDULED, ACTIVE, PUBLISHED_OPEN, CLOSED, RESULTS_PUBLISHED, ARCHIVED]
        assignmentId: { type: string }
        evaluatorId: { type: string }
        evaluateeId: { type: string }
        evaluatorRole: { type: string }
        stepType: { type: string, nullable: true }
        sectionId: { type: string, nullable: true }
        facultyId: { type: string, nullable: true }
        anonymityMode: { type: string, nullable: true }
        assignmentStatus: { type: string }
...